Optimize nushell

+++ home/sven/features/productivity/mail.nix
@@ -205,14 +205,22 @@ in {
   # if the agenix path changes.
   home.activation = {
     copySig = lib.hm.dag.entryAfter ["writeBoundary"] ''
-      rm -f ${leiderfischerSig}
-      rm -f ${effeffceeSig}
-      rm -f ${dgmSig}
-      rm -f ${taxdigitsSig}
-      cp ${config.age.secrets.leiderfischer.path} ${leiderfischerSig}
-      cp ${config.age.secrets.effeffcee.path} ${effeffceeSig}
-      cp ${config.age.secrets.dgm.path} ${dgmSig}
-      cp ${config.age.secrets.taxdigits.path} ${taxdigitsSig}
+      if [ -f ${leiderfischerSig} ]; then
+        rm -f ${leiderfischerSig}
+        cp ${config.age.secrets.leiderfischer.path} ${leiderfischerSig}
+      fi
+      if [ -f ${effeffceeSig} ]; then
+        rm -f ${effeffceeSig}
+        cp ${config.age.secrets.effeffcee.path} ${effeffceeSig}
+      fi
+      if [ -f ${dgmSig} ]; then
+        rm -f ${dgmSig}
+        cp ${config.age.secrets.dgm.path} ${dgmSig}
+      fi
+      if [ -f ${taxdigitsSig} ]; then
+        rm -f ${taxdigitsSig}
+        cp ${config.age.secrets.taxdigits.path} ${taxdigitsSig}
+      fi
     '';
   };
 }
